mini refactor: merge ConfirmEmail organism and page
