+++ src/components/pages/ConfirmEmail.scss
