+++ src/components/pages/ConfirmMail.jsx
@@ -1,9 +1,14 @@
 import React, { useEffect, useContext } from 'react'
+import { Link } from 'react-router-dom'
 
-import { ConfirmEmail as ConfirmEmailOrganism } from 'components/organisms/ConfirmEmail'
 import { useConfirmEmail } from 'hooks/useConfirmMail'
+import { useCounterLoop } from 'hooks/useCounterLoop'
 import { useProfile } from 'hooks/useProfile'
 import { SessionContext } from 'providers/SessionProvider'
+
+import { Logo } from 'Images'
+
+import classNames from './ConfirmEmail.scss'
 
 const useIfTruthy = fn => value => useEffect(() => {
   if (value)
@@ -22,11 +27,59 @@ export const ConfirmMail = ({ token }) => {
       setAccount({
         token: loginToken,
         email: profile.email,
+        issuer: profile.issuer,
       })
   }, [loginToken, profile])
 
   useLogIfTruthy(confirmEmailError)
   useLogIfTruthy(profileError)
 
-  return <ConfirmEmailOrganism loginToken={loginToken} error={confirmEmailError} />
+  return (
+    <section className={classNames.confirmEmail}>
+      <Header />
+      <Main loginToken={loginToken} error={confirmEmailError} />
+    </section>
+  )
 }
+
+const Header = () => (
+  <header>
+    <img src={Logo} />
+    <h1>Email Address Verification</h1>
+  </header>
+)
+
+const Main = ({ loginToken, error }) => (
+  <main>
+    {
+      !loginToken && !error
+        ? <InProgress />
+        : loginToken
+        ? <Success />
+        : <Failure reason={error} />
+    }
+  </main>
+)
+
+const InProgress = () => {
+  const counter = useCounterLoop()
+  return (
+    <section>
+      <h1>Verifying your email{'.'.repeat(counter + 1)}</h1>
+    </section>
+  )
+}
+
+const Success = () => (
+  <section>
+    <h1>Thank you! Your email is now verified :)</h1>
+    <Link to="/">Go to dashboard</Link>
+  </section>
+)
+
+const Failure = ({ reason }) => (
+  <section>
+    <h1>Oww, we could not verify your email address :(</h1>
+    <h2>{reason}</h2>
+  </section>
+)
